requirements.txt
paho-mqtt==1.6.1
pytest==7.3.1
